# Conflicts:
#	Ghostbait/Ghostbait/Ghostbait.vcxproj
#	Ghostbait/Ghostbait/Ghostbait.vcxproj.filters
#	Ghostbait/Ghostbait/Source Files/ObjectFactory.cpp

+++ Ghostbait/Omiracon/Genetics/Evolvable.cpp
@@ -7,7 +7,7 @@ namespace Omiracon {
 
 		void Evolvable::HandleMutationIteration(const MutationType& mutation, const size_t& curGen) {
 			traits.Mutate(mutation);
-			generation = curGen;
+			generation = (int)curGen;
 			performance.Reset();
 		}
 	}
